+++ src/shared-components/BrokerDetailsBreadcrumb/index.ts
@@ -1,0 +1,1 @@
+export * from './BrokerDetailsBreadcrumb';
