+++ src/shared-components/DropdownWithToggle/index.ts
@@ -1,0 +1,1 @@
+export * from './DropdownWithToggle';
